# backend/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "IRMS 后端 API（Gin + Gorm + Swagger）",
        "title": "IRMS API",
        "contact": {},
        "version": "0.1"
    },
    "basePath": "/api",
    "paths": {
...
        "/resource-group-members": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resource-groups"
                ],
                "summary": "列出主机组/服务组成员",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page (\u003e=1); invalid -\u003e 400",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (1~100); invalid -\u003e 400",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Group ID; invalid -\u003e 400",
                        "name": "group_id",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "host_group",
                            "service_group"
                        ],
                        "type": "string",
                        "description": "Group type; invalid -\u003e 400",
                        "name": "group_type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Deprecated: use group_id",
                        "name": "resource_group_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search keyword",
                        "name": "keyword",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtoresp.ResourceGroupMemberListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resource-groups"
                ],
                "summary": "添加资源组成员",
                "parameters": [
                    {
                        "description": "group_id/member_id/member_type 为主，resource_* 为兼容字段",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ResourceGroupMemberCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtoresp.EmptyResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resource-groups"
                ],
                "summary": "移除资源组成员",
                "parameters": [
                    {
                        "description": "group_id/member_id/member_type 为主，resource_* 为兼容字段",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ResourceGroupMemberDeleteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtoresp.EmptyResponse"
                        }
                    }
                }
            }
        },
        "/resource-groups": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resource-groups"
                ],
                "summary": "列出主机组/服务组",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page (\u003e=1); invalid -\u003e 400",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (1~100); invalid -\u003e 400",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search keyword (name like)",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "host_group",
                            "service_group"
                        ],
                        "type": "string",
                        "description": "Group type; invalid -\u003e 400",
                        "name": "group_type",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "host",
                            "service"
                        ],
                        "type": "string",
                        "description": "Deprecated: use group_type",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtoresp.ResourceGroupListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resource-groups"
                ],
                "summary": "创建资源组",
                "parameters": [
                    {
                        "description": "group_type 为主，type 为兼容字段",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ResourceGroupCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtoresp.IDResponse"
                        }
                    }
                }
            }
        },
        "/resource-groups/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resource-groups"
                ],
                "summary": "更新资源组",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Group ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "group_type 为主，type 为兼容字段",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ResourceGroupUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtoresp.EmptyResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resource-groups"
                ],
                "summary": "删除资源组",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Group ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtoresp.EmptyResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controller.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "details": {},
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
...
        "dtoresp.EmptyData": {
            "type": "object"
        },
        "dtoresp.EmptyResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {
                    "$ref": "#/definitions/dtoresp.EmptyData"
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
...
        "dtoresp.IDData": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "dtoresp.IDResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {
                    "$ref": "#/definitions/dtoresp.IDData"
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
...
        "dtoresp.Pagination": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "dtoresp.ResourceGroupListData": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/vo.ResourceGroupVO"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/dtoresp.Pagination"
                }
            }
        },
        "dtoresp.ResourceGroupListResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {
                    "$ref": "#/definitions/dtoresp.ResourceGroupListData"
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
        "dtoresp.ResourceGroupMemberListData": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/vo.ResourceGroupMemberVO"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/dtoresp.Pagination"
                }
            }
        },
        "dtoresp.ResourceGroupMemberListResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {
                    "$ref": "#/definitions/dtoresp.ResourceGroupMemberListData"
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
...
        "request.ResourceGroupCreateRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "group_type": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "request.ResourceGroupMemberCreateRequest": {
            "type": "object",
            "properties": {
                "group_id": {
                    "type": "integer"
                },
                "member_id": {
                    "type": "integer"
                },
                "member_name": {
                    "type": "string"
                },
                "member_type": {
                    "type": "string"
                },
                "resource_group_id": {
                    "type": "integer"
                },
                "resource_key": {
                    "type": "integer"
                },
                "resource_name": {
                    "type": "string"
                },
                "resource_type": {
                    "type": "string"
                }
            }
        },
        "request.ResourceGroupMemberDeleteRequest": {
            "type": "object",
            "properties": {
                "group_id": {
                    "type": "integer"
                },
                "member_id": {
                    "type": "integer"
                },
                "member_name": {
                    "type": "string"
                },
                "member_type": {
                    "type": "string"
                },
                "resource_group_id": {
                    "type": "integer"
                },
                "resource_key": {
                    "type": "integer"
                },
                "resource_name": {
                    "type": "string"
                },
                "resource_type": {
                    "type": "string"
                }
            }
        },
        "request.ResourceGroupUpdateRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "group_type": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "vo.ResourceGroupMemberVO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "group_id": {
                    "type": "integer"
                },
                "group_type": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "member_id": {
                    "type": "integer"
                },
                "member_name": {
                    "type": "string"
                },
                "member_type": {
                    "type": "string"
                }
            }
        },
        "vo.ResourceGroupVO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "group_type": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "member_count": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}